Reduce version_column_length for Doctrine migrations
Depending on the database version and settings, a key of length 1024 can
exceed the maximum length of an index. Reducing it to 255 is still more
than enough for any migration version, whilst ensuring it keeps within
the maximum index length (which is usually a multiple of the column
length).

Fixes #19